requirements.txt

altair==5.4.1
annotated-types==0.7.0
anyio==4.6.2.post1

absl-py==2.1.0
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
altair==5.4.1
appnope==0.1.4
asttokens==2.4.1
astunparse==1.6.3
async-timeout==4.0.3

attrs==24.2.0
Automat==24.8.1
beautifulsoup4==4.12.3
blinker==1.9.0

bs4==0.0.2


cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7

colorama==0.4.6
distro==1.9.0

comm==0.2.2
constantly==23.10.4
cryptography==43.0.3
cssselect==1.2.0
datasets==3.0.2
debugpy==1.8.7
decorator==5.1.1
defusedxml==0.7.1
dill==0.3.8

exceptiongroup==1.2.2
filelock==3.16.1

fsspec==2024.10.0
gitdb==4.0.11
GitPython==3.1.43
groq==0.12.0
h11==0.14.0
httpcore==1.0.7
httpx==0.27.2
huggingface-hub==0.26.2
idna==3.10
Jinja2==3.1.4
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1

flatbuffers==24.3.25
frozenlist==1.5.0
fsspec==2024.9.0
gast==0.6.0
gitdb==4.0.11
GitPython==3.1.43
google-pasta==0.2.0
grpcio==1.67.0
h11==0.14.0
h5py==3.12.1
huggingface-hub==0.26.1
hyperlink==21.0.0
idna==3.10
importlib_metadata==8.5.0
incremental==24.7.2
ipykernel==6.29.5
ipython==8.18.1
itemadapter==0.9.0
itemloaders==1.3.2
jedi==0.19.1
Jinja2==3.1.4
jmespath==1.0.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_client==8.6.3
jupyter_core==5.7.2
keras==3.6.0
libclang==18.1.1
lxml==5.3.0
Markdown==3.7

markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
mpmath==1.3.0

narwhals==1.14.1
networkx==3.4.2
numpy==2.1.2
outcome==1.3.0.post0
packaging==24.1
pandas==2.2.3
pillow==11.0.0
protobuf==5.28.3
pyarrow==18.0.0
pycparser==2.22
pydantic==2.10.0
pydantic_core==2.27.0
pydeck==0.9.1

multidict==6.1.0
multiprocess==0.70.16
namex==0.0.8
narwhals==1.14.1
nest-asyncio==1.6.0
networkx==3.2.1
numpy==1.26.4
opt_einsum==3.4.0
optree==0.13.0
outcome==1.3.0.post0
packaging==24.1
pandas==2.2.3
parsel==1.9.1
parso==0.8.4
pexpect==4.9.0
pillow==11.0.0
platformdirs==4.3.6
prompt_toolkit==3.0.48
propcache==0.2.0
Protego==0.3.1
protobuf==4.25.5
psutil==6.1.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==17.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydeck==0.9.1
PyDispatcher==2.0.7

Pygments==2.18.0
pyOpenSSL==24.2.1
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2

referencing==0.35.1
regex==2024.9.11
requests==2.32.3
rich==13.9.4
rpds-py==0.21.0
safetensors==0.4.5
scikit-learn==1.5.2
scipy==1.14.1
selenium==4.25.0
sentence-transformers==3.3.1

pyzmq==26.2.0
queuelib==1.7.0
referencing==0.35.1
regex==2024.9.11
requests==2.32.3
requests-file==2.1.0
rich==13.9.3
rpds-py==0.21.0
safetensors==0.4.5
Scrapy==2.11.2
selenium==4.25.0
sentencepiece==0.2.0
service-identity==24.2.0

six==1.16.0
smmap==5.0.1
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6

streamlit==1.40.1
sympy==1.13.1
tenacity==9.0.0
threadpoolctl==3.5.0
tokenizers==0.20.1
toml==0.10.2
torch==2.5.1
torchvision==0.20.1

stack-data==0.6.3
streamlit==1.40.1
sympy==1.13.1
tenacity==9.0.0
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorflow==2.16.2
tensorflow-io-gcs-filesystem==0.37.1
termcolor==2.5.0
tf_keras==2.16.0
tiktoken==0.8.0
tldextract==5.1.3
tokenizers==0.20.1
toml==0.10.2
tomli==2.1.0
torch==2.5.0
torchaudio==2.5.0
torchvision==0.20.0

tornado==6.4.1
tqdm==4.66.6
transformers==4.46.1
trio==0.27.0
trio-websocket==0.11.1
Twisted==24.10.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3

watchdog==6.0.0

w3lib==2.2.1
wcwidth==0.2.13

websocket-client==1.8.0
wsproto==1.2.0

xxhash==3.5.0
yarl==1.16.0
zipp==3.20.2
zope.interface==7.1.1

